Save text size preference in localStorage
